Add first version of PokerRoomsController (#8)
* Add @jest-mock/express

Because PokerRoomsController unit tests will require requests
and responses that can be quickly and consistently mocked

* Add PokerRoomsController create fn and tests

Initial version of create allows any user to create a table with
a name and optional buyIn, smallBlind, and bigBlind; no
input validation (tbd)

* Create common-api workspace

Because PokerRoomsController will need to return a json representation
the PokerRoom to clients, shared typings that are usable by both the cilent
and server will be useful. These shared files will live in a workspace
shared between client and server.

* Create Card.ts in common-api

To represent a Card returned from server to client

* Add Player to common-api

To represent the player (SelfPlayer) and opponents (OpponentPlayer)
returned from server to client

* Add Pot to common-api

To represent the Pots returned from server to client

* Add Table to common-api

To represent the Table returned from server to client

* Add missing dev packages to common-api

@types/node
@typescript-eslint/eslint-plugin
@typescript-eslint/parser
eslint
eslint-config-prettier
prettier

* Add index.ts to common-api

* Add main prop to common-api/package.json

* Update common-api tsconfig to emit type decs

Because other packages that depend on common-api need typings

* Add common-api as dependency in server

* Update common-api Player, common-api exports

Because common-api Player has two representaitons (Self and
Opponent), a more accurate typing reflects that holeCards
are only optional on Self (i.e. users can never see an opponent's
holeCards).

Because server will need to serialize value returned from
poker-engine's Player.legalActions() as an PlayerAction enum,
it needs to be exported for server's use.

* Add views/Card.ts

Because server needs a way to represent a player's view (i.e.
JSON representation returned to-be-returned client) of Table.

* Move holeCards to CommonPlayer

Because an opponent will need to have holeCards during showDowns, etc

* Remove Pick of "winners" from common-api Table

Because poker-engine's Table's winners are poker-engine's Players (i.e.
are a runtime only representation), winners need to be specified solely
as a common-api (i.e. serializable server-to-client) Player

* Move Card from views to views/player-views-of

Because these views are from the perspective of a player. Potential
future views may be from the perspective of an admin, a dealer, an
audience member, etc.

* Rename export in views/player-views-of

Because each file corresponds to a single view function, so no need for
named exports.

* Rename Card to ViewOfCard

Because want to follow convetion in other directories (e.g. controllers)
where file names have role appeneded. Did not go with CardView because
a view entails a subject (i.e. a viewer) and an object (i.e. a thing being
viewed)--CardView reads little more ambiguously wrt the subject/object
distinction.

* Update jsdoc for ViewOfCard

Because emphasis that this view is from the perspective of the player

* Add ViewOfPlayer

* Add ViewOfPot

* Add ViewOfTable

* Add Card.fixture

Because ViewOfCard will need to be tested on every possible Card
so adding a fixutre is appropriate

* Update CardSerlializer to use Card.fixture

* Remove unnecesasary import in CardSeralizer.test

* Add ViewOfCard.test

* Add Player.fixture

Because ViewOfPlayer.test can make use of fixutres for testing; these
same fixtures may potentially be used in other tests (e.g.
PayerSerializer.test)

* Change common-api Player to export Self, Opponent

Because ViewOfPlayer.test needs access to these types for proper typing

* Update Player.fixture descriptions

Because "last actor" is potentially ambiguous to testers, changed
description to "prior to current actor"

* Update ViewOfPlayer holeCard inclusion logic

Because players who have folded should not be able to view their own
holeCards.

* Update Player.fixture to throw when Player absent

Because Player.fixtures depend on Table.fixtures and Table.fixtures
could potentially be changed by a developer *independent of*
Player.fixtures, Player.fixtures should not assume field(s) containing
players (e.g. lastActor, currentActor, etc.) are necessarily defined;
instead Player.fixture should check for presence and throw Error in
exceptional cases.

* Add ViewOfPlayer.test.ts

With emphasis on visibility of holeCards and legalActions because this
layer will be responsible for preventing players from seeing opponents
secret information.

* Add Pot.fixture.ts

* Add ViewOfPot.test.ts

* Update ViewOfTable.ts

To conforom to default export norms of other views

* Add ViewOfTable.test.ts and associated snapshots

* Update ViewOfTable.ts to remove extra named export

* Upgrade @type/jest

To get mock.lastCall typing

* Update PokerRoomsController and test

To support get endpoint

* Update PokerRoomsController.ts with act handler

* Add act tests for PokerRoomsController.ts

* Update PokerRoomsController

Because get should return PokerRoom object that includes a
player's view of the table AND information about the room (i.e. name
and id) as well

Because PokerRoomsController needed JSDocs

* Update common-api to export req, res types

Because the client and the server should have identical definitions
of appropriate request and response body types, common-api
should supply an appropriate ReqBody and ResBody type for each server
endpoint. This means full typings for both were moved from the server
pacakge to the common-api package.

* Update server to use new common-api req, res typings

Because of previous commit, full typings for endpoint request and
response bodies were moved to common-api. As such, server needs
to import these typings and attach them to appropriate controllers,
views.