Update client dependencies to enable build
Because "eslintConfig": { "extends": [ "react-app" ] } in package.json
would result in the following error during build:
"Failed to load config "react-app" to extend from.",
eslint-config-react-app was added as an explicit devDependency.

Because adding eslint-config-react-app as a devDependency led to
react plugin having a conflict in eslint-config-react-app packages,
the following was added to .yarnrc.yml
packageExtensions:
  react-scripts@*:
    peerDependencies:
      eslint-config-react-app: "*"

See the following for explanation of above two:
https://github.com/facebook/create-react-app/issues/10463#issuecomment-1022751568

Becase @testing-library/jest-dom's types were recognized by
build command, @types/testing-library__jest-dom was added as
a devDependency.