Add deal action to server (#18)
* Update ViewOfPlayer to remove PokerEngine* naming

Because when file was written, Player and Routes.PokerRooms.Get.Player
had the same name (i.e. Player) and the disinction between 1) a Player
as represented by the engine and 2) a Player represented by the view
was necessary. With the namespacing of Routes.*, this is no longer
necessary.

* Add DEAL to common-api/PlayerAction, Act

* Add server/poker-engine/Utils, tests

Because @chevtek/poker-engine's Player class's legalActions method doesn't
return whether a given player can deal, support is needed for checking when
deal can be included in the legal actions a player can perform in the
viewOfPlayer.

* Update TableSerializer.test.ts.snap

Because Table.fixture was update in previous commit, snapshot tests for
TableSerializer.test need to be updated.

* Update ViewOfPlayer to include DEAL as legalAction

* Add DEAL action to PokerRoomsController, test