Update deserializers to use mutableContext
Deserializers used rely on Table.players as an exclusive list of all
Player objects that could appear (by refernce) in the rest of the
Table. That is, desializers would construct a fresh Player for each
serialized entry in Table.players. These deserialized Player elements
would then be used as references by other Player deserializers
(i.e. in Table.winners and Pot.eligiblePlayers).

However, in certain scenarios, a Player can exist in Table.winners or
Pot.eligiblePlayers (i.e. when a winner has stood up from the Table at
the end of a round).

This means that Player deserializers should have a more flexible
approach to addressing the construction/referencing issue.

The ultimate solution is to provide an optional mutableContext to
a deserializer which can allow Player deserializers
  1) check if the serialized Player references a previously constructed
    deserialized player and use the reference if present
  2) alternatively construct a new Player and save a reference in the
    mutableContext