Room creator page (#41)
* Move CreateRoom to components/create-room

* Remove formik, add react-hook-form dependency

* Add yup as peer, dev dependency to common-api

* Update Create to include reqBodySchema

Because client and server should be able to use common yup schema for
form and request validation, respectively

* Add common-api/jest.config

* Add Create.test

* Add msw with handler for post /api/rooms

Per guidance from https://kentcdodds.com/blog/stop-mocking-fetch, msw
will be used to mock response to network calls. Because CreateRoom
utilizes create mutation hook that sends post request to /api/rooms,
this is necessary for unit test.

* Update Create.ts reqBodySchema

Because when buyInLabel utilized a dash (i.e. "buy-in"), coversion w/
startCase in CreateRoom led to dash's removal (i.e. "Buy In"). In order
to align the error messages produced by the schema with the startCase,
dash was removed.

* Add CreateRoom

* Move create-room to room-creator

* Rename CreateRoom to RoomCreator

* Update client/package.json

Because it requires react-hook-form, msw, yup