Add page frame navigation (#42)
* Add navConfig

Because page title and nav drawer will depend on same human names for
paths, this allows a centralized place for specifying/computing the human
name

* Update navConfig

Because imports need organizing

* Add useNavTitle

Because title of page will depend on current URL path and navConfig, a
hook that returns this value will beuseful for future page title
components

* Add NavTitle

Because AppBar will need consistent title

* Update navConfig, dependencies

Because NavMenu tests will need to distinguish between title, humanName,
and parent of a given FlatNavConfigEntry.

* Add NavMenu

* Add navigation/index

Because other components should only need NavTitle or NavMenu--other
files, exports are for testing

* Update NavMenu.test

To use userEvent.setup()

* Update navigation/index

To export navConfig and flatNavConfig

* Update AppBar

So menu icon is hidden on medium and higher displays

* Add ResponsiveDrawer

* Update useNavTitle

To remove defaultTitle from options.

* Remove NavTitle

* Update AppBar

To receive title as prop without resorting to useNavTitle. Page title value
will be determined via a PageTitleProvider and associated hooks that
allow for overriding title in child components and fallback to navTitle
otherwise. This simplifies testing and separates concerns.

* Update navigation/index

To export useNavTitle hook.

* Add PageTitle

* Add PageFrame

* Move Room from src to src/components

* Update RoomSummary

To redirect to /room/:roomId instead of /rooms/:roomId

* Update RoomCreator

To remove title because title is now available in PageFrame

* Update RoomsList

To remove title becasue title is in PageFrame

* Update PageFrame

To include PageTitleProvider in place that usePageTitle can access

* Update App

To use PageFrame, ScrollToTop

* Update client/package.json

To import react-transition-group, css-mediaquery

* Add ScrollToTop